Update fight.js
Sorry for all the follow ups! Although my second concern is still there. and I'm wondering about the while statement (as the specifications say the game ends when one player's health falls below zero and not at zero, but if one is at zero then the while statement evaluates to false. I assumed that it was intended for the game to end when one player's health reached zero or below and that's what i checked for in the runRound function).